Synchronizer hooks. Week interval corected. MultiChart in url
