--- src/ducks/SANCharts/Chart/hook.js
@@ -1,10 +1,0 @@
-import { useEffect } from 'react'
-
-export const useResizeEffect = (clb, deps) =>
-  useEffect(() => {
-    window.addEventListener('resize', clb)
-
-    return () => {
-      window.removeEventListener('resize', clb)
-    }
-  }, deps)
